+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2020-05-XX
+
+- Remove `d3` as peerDependency
+- Add `d3-selection` and `d3-transition` as dependency
+- Remove export of `renderSelection`, but allow `render` to accept D3 `Selection` as selector
+
 ## [0.1.5] - 2020-04-29
 
 - Add `unpkg` to `package.json` for easier use with ObservableHQ